Pre-create bind mounts to ensure current user can write to them
When the directory for a bind mount doesn't exist on the host, the
Docker daemon creates it with the same user the daemon is running as.

If the Docker daemon is not running as the same user that the CLI is
running as, this can lead to a situation where the the container
(invoked with --user <current uuser UID:GID>) may not be able to write
to the bind mount as it was created with a different user.

For example, this can happen when the Docker daemon is running as root
but access to it doesn't require root (e.g. when added to the docker
group, or in WSL), or when a rootless Docker is running as a different
user to the current user.

The solution is to pre-create the directories with the current user the
CLI is running as, since the CLI user and the run container's user will
match.